# RemoteDoorControl/RPI/Nodejs/backend/app/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "RemoteDoorControl - SwaggerUI",
      "description": "",
      "version": "1.0"
    },
    "produces": ["application/json"],
    "paths": {
...
    "/api/location/Create": {
       "post": {
        "x-swagger-router-controller": "Main",
        "operationId": "impossible",
        "tags": ["Location"],
        "description": "add access location to DB and return all locations accessed",
        "parameters": [{
            "name": "geo1", 
            "in": "query", 
            "required": true, 
            "description": "longitude"
          },
          {
            "name": "geo2", 
            "in": "query", 
            "required": true, 
            "description": "latitude"
          }],
        "responses": {}
      }
     }
    }
 }